Split video dataset code into video_datasets.py, add PairedVideoLoaderConfig
Addresses review feedback from AnnaKwa on #1412:
- Move the video (temporal) batch/dataset classes and
  compute_calendar_features (formerly time_encoding.py) out of
  datasets.py into a dedicated video_datasets.py, since they're used
  independently of the existing spatial-downscaling data classes.
- Split PairedDataLoaderConfig's video-only args (n_timesteps,
  time_stride), clip_start_stride property, and build_video into a new
  PairedVideoLoaderConfig subclass, with docstrings for the new args, so
  callers can type-hint the more specific video config.
- Mark PairedDataLoaderConfig's methods @final now that it has an
  in-repo subclass, per repo convention.
- Rename test_video.py to test_video_datasets.py to match the module
  it tests.
